refactor: consolidate integration prompts into per-integration builders
Rename buildIntegrationContext to buildIntegrationPrompt and the file from
integration-context.ts to integration-prompt.ts to match the existing naming
convention (buildAgentIdentityPrompt, buildAgentToolsPrompt).

Create per-integration builder functions (buildVoiceChatWorkerPrompt,
buildSlackPrompt, buildPhonePrompt, buildTelegramPrompt, buildGitHubPrompt)
that encapsulate prompt assembly, simplifying all handler call sites.

Move VOICE_CHAT_WORKER_PROMPT from session-service.ts into the integration
prompt module where it belongs.

Closes #8696
