+++ turbo/apps/web/src/lib/zero/email/handlers/inbound-trigger.ts
@@ -11,7 +11,7 @@ import {
   type HandlerResult,
 } from "./shared";
 import { createZeroRun } from "../../zero-run-service";
-import { buildIntegrationContext } from "../../integration-context";
+import { buildIntegrationPrompt } from "../../integration-prompt";
 import { generateCallbackSecret, getApiUrl } from "../../../infra/callback";
 import { getUserIdByEmail } from "../../../auth/get-user-id-by-email";
 import { getOrgIdBySlug } from "../../../auth/org-cache";
@@ -224,7 +224,7 @@ export async function handleInboundEmailTrigger(
   ];
 
   // 12. Create run with integration context as system prompt
-  const appendSystemPrompt = buildIntegrationContext("Email");
+  const appendSystemPrompt = buildIntegrationPrompt("Email");
   const result = await createZeroRun({
     userId,
     prompt,
