--- turbo/apps/web/src/lib/zero/integration-context.ts
@@ -1,85 +1,0 @@
-/**
- * The platform (integration) an agent run originates from.
- * Used to inject context so the agent knows which channel it is operating in.
- */
-type IntegrationPlatform =
-  | "Email"
-  | "GitHub"
-  | "Phone"
-  | "Schedule"
-  | "Slack"
-  | "Telegram"
-  | "Voice-Chat";
-
-/**
- * Build the integration context header prepended to agent run prompts.
- */
-export function buildIntegrationContext(
-  platform: IntegrationPlatform,
-  options?: {
-    botUserId?: string;
-    channelId?: string;
-    channelType?: "channel" | "dm" | "group_dm";
-    threadId?: string;
-    scheduleDescription?: string;
-    triggerType?: string;
-  },
-): string {
-  let context = `# Current Integration\nYou are currently running inside: ${platform}`;
-  if (options?.botUserId) {
-    context += `\nYour bot user ID: ${options.botUserId}`;
-  }
-  if (options?.channelId) {
-    context += `\nChannel ID: ${options.channelId}`;
-  }
-  if (options?.channelType) {
-    const typeLabel =
-      options.channelType === "dm"
-        ? "Direct message"
-        : options.channelType === "group_dm"
-          ? "Group direct message"
-          : "Channel";
-    context += `\nChannel type: ${typeLabel}`;
-  }
-  if (options?.threadId) {
-    context += `\nThread ID: ${options.threadId}`;
-  }
-  if (options?.scheduleDescription) {
-    context += `\nSchedule description: ${options.scheduleDescription}`;
-  }
-  if (options?.triggerType) {
-    context += `\nTrigger type: ${options.triggerType}`;
-  }
-  return context;
-}
-
-export interface UserInfoOptions {
-  name?: string;
-  email?: string;
-  timezone?: string;
-  slackDisplayName?: string;
-  slackUserId?: string;
-}
-
-/**
- * Build the user info section for agent system prompts.
- */
-export function buildUserInfo(options: UserInfoOptions): string {
-  const lines: string[] = [];
-  if (options.name) {
-    lines.push(`Name: ${options.name}`);
-  }
-  if (options.email) {
-    lines.push(`Email: ${options.email}`);
-  }
-  if (options.timezone) {
-    lines.push(`Timezone: ${options.timezone}`);
-  }
-  if (options.slackDisplayName) {
-    lines.push(`Slack display name: ${options.slackDisplayName}`);
-  }
-  if (options.slackUserId) {
-    lines.push(`Slack user ID: ${options.slackUserId}`);
-  }
-  return `# Current User Info\n${lines.join("\n")}`;
-}
